chore: Move legacy shadow nodes and component descriptors to legacy directory
