+++ common/cpp/react/renderer/components/rnscreens/legacy/RNSScreenComponentDescriptor.h
@@ -7,10 +7,10 @@
 #endif // ANDROID
 #include <react/debug/react_native_assert.h>
 #include <react/renderer/components/rnscreens/Props.h>
-#include <react/renderer/components/rnscreens/utils/RectUtil.h>
 #include <react/renderer/components/root/RootShadowNode.h>
 #include <react/renderer/core/ConcreteComponentDescriptor.h>
 #include "RNSScreenShadowNode.h"
+#include "utils/RectUtil.h"
 
 namespace facebook {
 namespace react {
