+++ common/cpp/react/renderer/components/rnscreens/legacy/RNSScreenStackHeaderConfigComponentDescriptor.h
@@ -5,9 +5,9 @@
 #endif
 #include <react/debug/react_native_assert.h>
 #include <react/renderer/components/rnscreens/Props.h>
-#include <react/renderer/components/rnscreens/utils/RectUtil.h>
 #include <react/renderer/core/ConcreteComponentDescriptor.h>
 #include "RNSScreenStackHeaderConfigShadowNode.h"
+#include "utils/RectUtil.h"
 
 namespace facebook::react {
