+++ common/cpp/react/renderer/components/rnscreens/legacy/RNSScreenStackHeaderSubviewState.h
@@ -9,7 +9,7 @@
 #include <react/renderer/mapbuffer/MapBufferBuilder.h>
 #endif
 
-#include "FrameCorrectionModes.h"
+#include "../FrameCorrectionModes.h"
 
 namespace facebook::react {
