+++ cpp/legacy/RNSScreenRemovalListener.cpp
