+++ cpp/legacy/RNSScreenRemovalListener.h
